requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.4
django-anymail==8.2
django-cors-headers==3.5.0
django-heroku==0.0.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
gunicorn==19.0.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mpmath==1.2.1
openapi-codec==1.3.2
Pillow==8.0.1
psycopg2-binary==2.8.6
pyactiveresource==1.0.1
pytz==2020.4
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
stripe==2.56.0
sympy==1.7.1
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
